performance: optimize by protobuf context-type.
